+++ designer/loadable-config-api/src/main/scala/pl/touk/nussknacker/ui/loadableconfig/DesignerRootConfig.scala
@@ -1,0 +1,16 @@
+package pl.touk.nussknacker.ui.loadableconfig
+
+import com.typesafe.config.Config
+import pl.touk.nussknacker.engine.ConfigWithUnresolvedVersion
+
+// TODO: We should extract a class for all configuration options that should be available to designer instead of returning raw hocon config.
+//       Thanks to that it will be easier to split processing type config from rest of configs and use this interface programmatically
+final case class DesignerRootConfig(rawConfig: ConfigWithUnresolvedVersion)
+
+object DesignerRootConfig {
+
+  def from(config: Config): DesignerRootConfig = {
+    DesignerRootConfig(ConfigWithUnresolvedVersion(config))
+  }
+
+}
