+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/server/RouteProvider.scala
@@ -2,7 +2,7 @@ package pl.touk.nussknacker.ui.server
 
 import akka.http.scaladsl.server.Route
 import cats.effect.{IO, Resource}
-import pl.touk.nussknacker.ui.config.DesignerRootConfig
+import pl.touk.nussknacker.ui.loadableconfig.DesignerRootConfig
 
 trait RouteProvider[R <: Route] {
