+++ extensions-api/src/main/scala/pl/touk/nussknacker/engine/ConfigWithUnresolvedVersion.scala
@@ -2,10 +2,29 @@ package pl.touk.nussknacker.engine
 
 import com.typesafe.config.{Config, ConfigFactory, ConfigResolveOptions}
 
+import scala.jdk.CollectionConverters._
+
 case class ConfigWithUnresolvedVersion private (withUnresolvedEnvVariables: Config, resolved: Config) {
 
   def getConfig(path: String): ConfigWithUnresolvedVersion = {
     ConfigWithUnresolvedVersion(withUnresolvedEnvVariables.getConfig(path), resolved.getConfig(path))
+  }
+
+  def readMap(path: String): Option[Map[String, ConfigWithUnresolvedVersion]] = {
+    if (resolved.hasPath(path)) {
+      val nestedConfig = getConfig(path)
+      Some(
+        nestedConfig.resolved
+          .root()
+          .entrySet()
+          .asScala
+          .map(_.getKey)
+          .map { key => key -> nestedConfig.getConfig(key) }
+          .toMap
+      )
+    } else {
+      None
+    }
   }
 
 }
